insurance_claim__agentic_ai/requirements.txt
# Web framework
# flask==3.1.2
# gunicorn==21.2.0

# # Database
# #sqlalchemy==2.0.25
# pip install psycopg2-binary

# # Security
# #pyjwt==2.8.0
# PyJWT-2.11.0


# # Machine Learning
# xgboost==2.0.3
# scikit-learn==1.3.2
# numpy==1.26.4
# pandas==2.2.1


# # Utilities
# python-dotenv==1.0.1


Flask==2.3.3
#PyJWT==2.8.0
Werkzeug==2.3.7
psycopg2-binary==2.9.6
numpy==1.26.0
#numpy==1.26.4
pandas==2.1.1
xgboost==2.0.3
scikit-learn==1.3.2
Jinja2==3.1.3
joblib==1.3.2
python-dotenv==1.0.1
sqlalchemy==2.0.25
# pip install psycopg2-binary
#pyjwt==2.8.0
PyJWT==2.11.0
#openai==1.30.1
google-generativeai


